+++ fastdeploy/logger/deterministic_logger.py
@@ -20,7 +20,16 @@ import time
 import numpy as np
 import paddle
 
-from fastdeploy.utils import parse_choice_id
+_CHOICE_SEPARATOR = "::n::"
+
+
+def _parse_choice_id(compound_id: str) -> tuple:
+    """Parse an internal request ID into (base_request_id, choice_index)."""
+    if _CHOICE_SEPARATOR in compound_id:
+        base, idx = compound_id.rsplit(_CHOICE_SEPARATOR, 1)
+        return base, int(idx)
+    return compound_id, None
+
 
 det_logger = logging.getLogger("fastdeploy.deterministic")
 
@@ -130,7 +139,7 @@ class DeterministicLogger:
         current_run_id = None
         for req in model_forward_batch or []:
             if req is not None:
-                _, index = parse_choice_id(req.request_id)
+                _, index = _parse_choice_id(req.request_id)
                 if index is not None:
                     current_run_id = str(index)
                     break
